agentcore-strands-db-mcp-assistant/requirements.txt
strands-agents
strands-agents-tools
typing-extensions
boto3>=1.34.0
bedrock-agentcore>=0.0.8
bedrock-agentcore-starter-toolkit
botocore>=1.34.0
mcp